+++ di-df1/lib/Di/Df1.hs
@@ -396,7 +396,7 @@ debug' natSTM di = Di.log' natSTM di Df1.Debug . Df1.message
 df1 :: Di.LineRenderer Df1.Level Df1.Path Df1.Message
 {-# INLINE df1 #-}
 df1 = Di.LineRendererUtf8 (\x ->
-  if x then Df1.Render.logColor . fromDiLog
+  if x then Df1.Render.logColorANSI . fromDiLog
        else Df1.Render.log . fromDiLog)
 
 --------------------------------------------------------------------------------