+++ df1/CHANGELOG.md
@@ -1,3 +1,12 @@
+# Version ???
+
+* COMPILER ASSISTED BREAKING CHANGE: Change functions names `render` to `log`,
+  `renderColor` to `logColor` and export from  `Df1.Render` instead of `Df1`.
+  
+* Exported `key`, `message`, `iso8601`, `segment` and `value` from module 
+  `Df1.Render`.
+
+
 # Version 0.3.2
 
 * Added `ToValue` instances for common types like `Int`, `Bool`, etc.